+++ src/tests/test_lsp/mod.rs
@@ -3,6 +3,7 @@ pub mod bench_completion;
 pub mod lsp_client;
 pub mod test_code_action;
 pub mod test_completion;
+pub mod test_diagnostics;
 pub mod test_goto_definition;
 pub mod test_hover;
 pub mod test_references;